Udacity FSND Linux Server Configuration Walkthough
This is the project is in fulfillment of the Udacity Fullstack
Nanodegree program, and outlines how to set up a remote Linux Server
using Amazon AWS Lightsail, and deploy a Flask application. The
application hosted on this virtual server is the Item Catalog project,
which was also in fulfillment of the Udacity Fullstack Nanodegree
program.